requirements.txt
# Core dependencies
asyncio
logging
typing

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
alembic>=1.12.0

# Speech Recognition (Whisper) - Local
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Speech Recognition - Cloud APIs
deepgram-sdk>=3.0.0          # Deepgram STT
azure-cognitiveservices-speech>=1.31.0  # Azure Speech Services
google-cloud-speech>=2.21.0   # Google Cloud STT

# Text-to-Speech - Local
TTS>=0.17.0                   # Coqui TTS
mimic3-tts>=0.2.3            # Mimic3 TTS

# Text-to-Speech - Cloud APIs
elevenlabs>=0.2.0            # ElevenLabs TTS
google-cloud-texttospeech>=2.14.0  # Google Cloud TTS
# Azure TTS uses same package as Azure STT (azure-cognitiveservices-speech)

# Audio processing
pyaudio>=0.2.11
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.0

# Machine Learning & AI
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
transformers>=4.35.0
sentiment-analysis>=0.1.0

# Emotion Recognition
fer>=22.5.1
opencv-python>=4.8.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# HTTP and API
requests>=2.31.0
aiohttp>=3.9.0

# State Machine
transitions>=0.9.0

# Logging and monitoring
loguru>=0.7.0

# Telephony Integration (Asterisk)
# Uses standard library: socket, threading, queue, asyncio
# No additional dependencies required for basic AMI integration

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0